requirements.txt
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0

